feat: 함께해요 게시글 마감 기능 구현
