+++ app/src/main/java/com/umc/ttoklip/data/model/town/PatchCartStatusRequest.kt
@@ -1,0 +1,5 @@
+package com.umc.ttoklip.data.model.town
+
+data class PatchCartStatusRequest(
+    val status: String
+)
