+++ app/src/main/java/com/umc/ttoklip/data/repository/town/ReadTogetherRepository.kt
@@ -4,6 +4,7 @@ import com.umc.ttoklip.data.model.CommonResponse
 import com.umc.ttoklip.data.model.StandardResponse
 import com.umc.ttoklip.data.model.town.CommentResponse
 import com.umc.ttoklip.data.model.town.CreateCommentRequest
+import com.umc.ttoklip.data.model.town.PatchCartStatusRequest
 import com.umc.ttoklip.data.model.town.ReportRequest
 import com.umc.ttoklip.data.model.town.ViewTogetherResponse
 import com.umc.ttoklip.module.NetworkResult
@@ -19,4 +20,6 @@ interface ReadTogetherRepository {
     suspend fun cancelTogether(postId: Long): NetworkResult<CommonResponse>
 
     suspend fun fetchParticipantsCount(postId: Long): NetworkResult<CommonResponse>
+
+    suspend fun patchPostStatus(postId: Long, request: PatchCartStatusRequest): NetworkResult<CommonResponse>
 }
