+++ app/src/main/java/com/umc/ttoklip/data/repository/town/ReadTogetherRepositoryImpl.kt
@@ -6,6 +6,7 @@ import com.umc.ttoklip.data.model.ResponseBody
 import com.umc.ttoklip.data.model.StandardResponse
 import com.umc.ttoklip.data.model.town.CommentResponse
 import com.umc.ttoklip.data.model.town.CreateCommentRequest
+import com.umc.ttoklip.data.model.town.PatchCartStatusRequest
 import com.umc.ttoklip.data.model.town.ReportRequest
 import com.umc.ttoklip.data.model.town.ViewTogetherResponse
 import com.umc.ttoklip.module.NetworkResult
@@ -68,5 +69,9 @@ class ReadTogetherRepositoryImpl @Inject constructor(private val api: ReadTogeth
 
     override suspend fun fetchParticipantsCount(postId: Long): NetworkResult<CommonResponse> {
         return handleApi({api.fetchParticipantsCount(postId)}) {response: ResponseBody<CommonResponse> -> response.result}
+    }
+
+    override suspend fun patchPostStatus(postId: Long, request: PatchCartStatusRequest): NetworkResult<CommonResponse> {
+        return handleApi({api.patchPostStatus(postId, request)}) {response: ResponseBody<CommonResponse> -> response.result}
     }
 }
