+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/communication/read/ReadCommunicationViewModelImpl.kt
@@ -191,4 +191,5 @@ class ReadCommunicationViewModelImpl @Inject constructor(
             }
         }
     }
+
 }
