+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/together/read/ReadTogetherViewModel.kt
@@ -32,4 +32,6 @@ interface ReadTogetherViewModel {
     fun cancelTogether()
 
     fun fetchParticipantsCount()
+
+    fun patchPostStatus(status: String)
 }
